# hash/maphash: flush the buffer during Sum64