@kixelated 2022-01-16
Ian pointed out that this will not work, because the internal has does not work on partial inputs. I wrote a quick unit test to confirm.

```golang
func TestSumSequence(t *testing.T) {
  seed := maphash.MakeSeed()
  
  h1 := new(maphash.Hash)
  h1.SetSeed(seed)
  h1.WriteString("foo")
  h1.WriteString("bar")
  s1 := h1.Sum64()
  
  h2 := new(maphash.Hash)
  h2.SetSeed(seed)
  h2.WriteString("foo")
  h2.Sum64()
  h2.WriteString("bar")
  s2 := h2.Sum64()

  if s1 != s2 {
    t.Error("hashes don't match")
  }
}
```

Oh well.